lesson33 added ValidateService + visual fix
