+++ Lesson33/src/main/java/by/tms/controller/impl/CreateUserInputPageCommandController.java
@@ -7,7 +7,6 @@ import static by.tms.utils.Constants.RequestParameters.LOGIN;
 import static by.tms.utils.Constants.RequestParameters.VERIFY_PASSWORD;
 import static by.tms.utils.ControllerUtils.getUser;
 import static by.tms.utils.ServletUtils.saveUserSession;
-import static by.tms.utils.ValidatorUtils.isVerifyUserData;
 
 import by.tms.controller.CommandController;
 import by.tms.exception.CommandException;
@@ -15,6 +14,7 @@ import by.tms.model.Inject;
 import by.tms.model.PagesPath;
 import by.tms.model.User;
 import by.tms.service.UserService;
+import by.tms.service.ValidateService;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.servlet.http.HttpServletRequest;
@@ -25,14 +25,15 @@ public class CreateUserInputPageCommandController implements CommandController {
 
     @Inject
     private UserService userService;
+    @Inject
+    private ValidateService validateService;
 
     @Override
     public PagesPath execute(HttpServletRequest request) throws CommandException {
         String login = request.getParameter(LOGIN);
         String verifyPassword = request.getParameter(VERIFY_PASSWORD);
         User user = getUser(request, login);
-        List<String> errorMessages = isVerifyUserData(user);
-        checkUniqueUser(verifyPassword, user, errorMessages);
+        List<String> errorMessages = validateService.getValidationErrorMessage(user, verifyPassword);
         PagesPath path;
         if (errorMessages.isEmpty()) {
             userService.addUser(user);
@@ -45,18 +46,5 @@ public class CreateUserInputPageCommandController implements CommandController {
             path = FAIL_REGISTER_PAGE;
         }
         return path;
-    }
-
-    private void checkUniqueUser(String verifyPassword, User user, List<String> errorMessages) {
-        if (!isNewUserVerify(user.getLogin(), user.getPassword(), verifyPassword)) {
-            errorMessages.add("This user already exist");
-        }
-    }
-
-    private boolean isNewUserVerify(String login, String password, String verifyPassword) {
-        if (userService.getUserByLogin(login) == null) {
-            return password.equals(verifyPassword);
-        }
-        return false;
     }
 }
