+++ Lesson33/src/main/java/by/tms/service/ValidateService.java
@@ -1,0 +1,9 @@
+package by.tms.service;
+
+import by.tms.model.User;
+import java.util.List;
+
+public interface ValidateService {
+
+    List<String> getValidationErrorMessage(User user, String verifyPassword);
+}
