+++ Lesson33/src/main/java/by/tms/utils/ValidatorUtils.java
@@ -1,10 +1,7 @@
 package by.tms.utils;
 
-import by.tms.model.User;
 import java.time.LocalDate;
 import java.time.Period;
-import java.util.ArrayList;
-import java.util.List;
 import lombok.experimental.UtilityClass;
 
 @UtilityClass
@@ -24,22 +21,5 @@ public class ValidatorUtils {
 
     public static boolean isLoginPasswordVerify(String login, String password) {
         return login.matches("[a-zA-Z0-9]{4,30}") && password.matches("[a-zA-Z0-9]{4,30}");
-    }
-
-    public static List<String> isVerifyUserData(User user) {
-        List<String> errorMessages = new ArrayList<>();
-        if (!isLoginPasswordVerify(user.getLogin(), user.getPassword())) {
-            errorMessages.add("Incorrect login or password");
-        }
-        if (!isNameSurnameVerify(user.getName(), user.getSurname())) {
-            errorMessages.add("Incorrect name or surname");
-        }
-        if (!isEmailVerify(user.getEmail())) {
-            errorMessages.add("Incorrect email");
-        }
-        if (!isAgeVerify(user.getBirthday())) {
-            errorMessages.add("Registration is available from the age of 18");
-        }
-        return errorMessages;
     }
 }
